Make fixes after rebase

+++ mlir/include/air/Util/Util.h
@@ -30,7 +30,7 @@ func::FuncOp getMangledFunction(ModuleOp module, std::string fnName,
 uint64_t getTensorVolume(const ShapedType ty);
 uint64_t getTensorVolume(const Type ty);
 std::string getElementTypeAsString(const mlir::Type ty);
-unsigned getElementSizeInBytes(const mlir::Type ty);
+uint64_t getElementSizeInBytes(const mlir::Type ty);
 
 // Get the parent scf.for op of an iter_arg
 scf::ForOp getForRegionIterArgsOwner(Value val);
