+++ mlir/lib/Util/Util.cpp
@@ -158,7 +158,7 @@ std::string air::getElementTypeAsString(const mlir::Type ty) {
 }
 
 // An incomplete lookup table of common data types
-unsigned air::getElementSizeInBytes(const mlir::Type ty) {
+uint64_t air::getElementSizeInBytes(const mlir::Type ty) {
   auto typeAsString = getElementTypeAsString(ty);
   if (typeAsString == "i32")
     return 4;
